move Gui to Gui_generate.py
