+++ MAIN.py
@@ -1,3 +1,11 @@
+from tkinter import Tk
 
 from import_checker import *
 main()
+
+from gui_tree_import_checker import *
+
+from gui_generate import *
+root = Tk()
+app = Gui(root=root)
+app.mainloop()
